+++ databasez/dialects/jdbc.py
@@ -9,11 +9,14 @@ from sqlalchemy.connectors.asyncio import (
 from sqlalchemy.engine.default import DefaultDialect
 from sqlalchemy.pool import AsyncAdaptedQueuePool
 from sqlalchemy.util.concurrency import await_only
+from sqlalchemy_utils.functions.orm import quote
+from sqlalchemy.sql import text
 
 from databasez.utils import AsyncWrapper
 
 if typing.TYPE_CHECKING:
     from sqlalchemy import URL
+    from sqlalchemy.base import Connection
     from sqlalchemy.engine.interfaces import ConnectArgsType
 
 
@@ -27,6 +30,7 @@ class JDBC_dialect(DefaultDialect):
     """
 
     driver = "jdbc"
+    supports_statement_cache = True
 
     is_async = True
 
@@ -34,26 +38,29 @@ class JDBC_dialect(DefaultDialect):
         self,
         json_serializer: typing.Any = None,
         json_deserializer: typing.Any = None,
-        **kwargs,
-    ):
+        **kwargs: typing.Any,
+    ) -> None:
         super().__init__(**kwargs)
 
     def create_connect_args(self, url: "URL") -> "ConnectArgsType":
-        jdbc_dsn_driver: str = url.query["jdbc_dsn_driver"]
-        jdbc_driver: typing.Optional[str] = url.query.get("jdbc_driver")
+        jdbc_dsn_driver: str = typing.cast(str, url.query["jdbc_dsn_driver"])
+        jdbc_driver: typing.Optional[str] = typing.cast(typing.Optional[str], url.query.get("jdbc_driver"))
         driver_args: typing.Any = url.query.get("jdbc_driver_args")
         if driver_args:
             driver_args = orjson.loads(driver_args)
         dsn: str = url.difference_update_query(
-            ("jdbc_driver", "jdbc_driver_args")
+            ("jdbc_driver", "jdbc_driver_args", "jdbc_dsn_driver")
         ).render_as_string(hide_password=False)
-        dsn = dsn.replace("jdbc:", f"jdbc:{jdbc_dsn_driver}", 1)
+        dsn = dsn.replace("jdbc://", f"jdbc:{jdbc_dsn_driver}:", 1)
 
         return (dsn,), {"driver_args": driver_args, "driver": jdbc_driver}
 
     def connect(self, *arg: typing.Any, **kw: typing.Any) -> AsyncAdapt_adbapi2_connection:
         creator_fn = AsyncWrapper(
-            self.loaded_dbapi, pool=ThreadPoolExecutor(1, thread_name_prefix="jpype")
+            self.loaded_dbapi,
+            pool=ThreadPoolExecutor(1, thread_name_prefix="jpype"),
+            stringify_exceptions=True,
+            exclude_attrs={"connect": {"cursor": True}},
         ).connect
 
         creator_fn = kw.pop("async_creator_fn", creator_fn)
@@ -63,8 +70,27 @@ class JDBC_dialect(DefaultDialect):
             await_only(creator_fn(*arg, **kw)),
         )
 
+    def has_table(
+        self,
+        connection: "Connection",
+        table_name: str,
+        schema: typing.Optional[str] = None,
+        **kw: typing.Any,
+    ) -> bool:
+        stmt = text(f"select * from '{quote(connection, table_name)}' LIMIT 1")
+        try:
+            connection.execute(stmt)
+            return True
+        except Exception as exc:
+            return False
+
+    def get_isolation_level(
+        self, dbapi_connection: typing.Any
+    ) -> typing.Any:
+        return None
+
     @classmethod
-    def get_pool_class(cls, url):
+    def get_pool_class(cls, url: "URL") -> typing.Any:
         return AsyncAdaptedQueuePool
 
     @classmethod