+++ databasez/overwrites/dbapi2.py
@@ -1,16 +1,25 @@
 import typing
 
-from databasez.sqlalchemy import SQLAlchemyDatabase
+from databasez.sqlalchemy import SQLAlchemyDatabase, SQLAlchemyTransaction
 
 if typing.TYPE_CHECKING:
     from databasez.core.databaseurl import DatabaseURL
 
 
+class Transaction(SQLAlchemyTransaction):
+    def get_default_transaction_isolation_level(
+        self, is_root: bool, **extra_options: typing.Any
+    ) -> typing.Optional[str]:
+        return None
+
+
 class Database(SQLAlchemyDatabase):
+    default_isolation_level = None
+
     def extract_options(
         self,
         database_url: "DatabaseURL",
-        **options: typing.Dict[str, typing.Any],
+        **options: typing.Any,
     ) -> typing.Tuple["DatabaseURL", typing.Dict[str, typing.Any]]:
         database_url_new, options = super().extract_options(database_url, **options)
         new_query_options = dict(database_url.options)